src/prescriptions.api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
azure-servicebus==7.12.3
redis==5.2.0
scikit-learn==1.5.2
numpy==2.1.3
joblib==1.4.2
openai==1.54.4
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
structlog==24.4.0
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
